# gateway-service/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Gateway Service API
  version: 1.0.0
  description: API for statements and documents operations and administrative actions
servers:
  - url: http://localhost:8084
    description: Local server

paths:
  /gateway/statement:
    post:
      tags:
        - Statement
      operationId: getOffers
      summary: Getting loan offers
      description: Gets loan offers with all possible loan terms
      requestBody:
        description: Data for getting loan offers
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanStatementRequestDto'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoanOfferDto'
              examples:
                fourOffers:
                  summary: Example with 4 loan offers
                  value:
                    - statementId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                      requestedAmount: 100000
                      totalAmount: 150000
                      term: 36
                      monthlyPayment: 10000
                      rate: 20
                      isInsuranceEnabled: false
                      isSalaryClient: false
                    - statementId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                      requestedAmount: 100000
                      totalAmount: 140000
                      term: 36
                      monthlyPayment: 8000
                      rate: 16
                      isInsuranceEnabled: true
                      isSalaryClient: false
                    - statementId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                      requestedAmount: 100000
                      totalAmount: 130000
                      term: 36
                      monthlyPayment: 6000
                      rate: 14
                      isInsuranceEnabled: false
                      isSalaryClient: true
                    - statementId: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                      requestedAmount: 100000
                      totalAmount: 120000
                      term: 36
                      monthlyPayment: 4000
                      rate: 12
                      isInsuranceEnabled: false
                      isSalaryClient: false
        '400':
          description: Data did not pass prescoring
        '424':
          description: Failed to get info from statement service

  /gateway/offer:
    post:
      tags:
        - Statement
      operationId: selectOffer
      summary: Choosing one of the offers
      description: Accepts one offer
      requestBody:
        description: Data for choosing one of the loan offer
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanOfferDto'
      responses:
        '200':
          description: Successful operation
        '404':
          description: Statement not found
        '424':
          description: Failed to get info from statement service

  /gateway/calculate/{statementId}:
    post:
      tags:
        - Statement
      operationId: finishRegistration
      summary: Completion of registration
      description: Finishes registration and makes a full calculation of the loan
      parameters:
        - name: statementId
          in: path
          description: Unique id of loan statement
          required: true
          schema:
            type: string
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      requestBody:
        description: Data for completion of registration
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FinishRegistrationRequestDto'
      responses:
        '200':
          description: Successful operation
        '404':
          description: Statement not found
        '409':
          description: Document is already signed
        '424':
          description: Failed to get info from deal service

  /gateway/document/{statementId}/send:
    post:
      tags:
        - Document
      operationId: sendDocumentsRequest
      summary: Sending documents request
      description: Sends a request to send documents
      parameters:
        - name: statementId
          in: path
          description: Unique id of loan statement
          required: true
          schema:
            type: string
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      responses:
        '200':
          description: Successful operation
        '404':
          description: Statement not found
        '409':
          description: Document is already signed
        '424':
          description: Failed to get info from deal service

  /gateway/document/{statementId}/sign:
    post:
      tags:
        - Document
      operationId: signDocumentsRequest
      summary: Signing documents request
      description: Sends a request to sign documents
      parameters:
        - name: statementId
          in: path
          description: Unique id of loan statement
          required: true
          schema:
            type: string
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      responses:
        '200':
          description: Successful operation
        '404':
          description: Statement not found
        '409':
          description: Document is already signed
        '424':
          description: Failed to get info from deal service

  /gateway/document/{statementId}/code:
    post:
      tags:
        - Document
      operationId: signDocuments
      summary: Signing documents
      description: Signs documents and issues a credit
      parameters:
        - name: statementId
          in: path
          description: Unique id of loan statement
          required: true
          schema:
            type: string
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        - name: code
          in: query
          description: Code for signing documents
          required: true
          schema:
            type: string
          example: "A7B9Z2"
      responses:
        '200':
          description: Successful operation
        '400':
          description: Invalid code
        '404':
          description: Statement not found
        '409':
          description: Document is already signed
        '424':
          description: Failed to get info from deal service

  /gateway/admin/statement/{statementId}:
    get:
      tags:
        - Admin
      operationId: getStatement
      summary: Getting statement
      description: Gets statement information by id
      parameters:
        - name: statementId
          in: path
          description: Unique id of statement
          required: true
          schema:
            type: string
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatementDto'
        '404':
          description: Statement not found
        '424':
          description: Failed to get info from deal service

  /gateway/admin/statement:
    get:
      tags:
        - Admin
      operationId: getStatements
      summary: Getting statements
      description: Gets all statements
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StatementDto'
        '424':
          description: Failed to get info from deal service

components:
  schemas:
    LoanStatementRequestDto:
      type: object
      description: Data for creating loan offers
      required:
        - amount
        - term
        - firstName
        - lastName
        - email
        - birthDate
        - passportSeries
        - passportNumber
      properties:
        amount:
          type: number
          format: bigdecimal
          minimum: 10000
          maximum: 100000000
          description: Requested loan amount
          example: 100000
        term:
          type: integer
          minimum: 3
          maximum: 360
          description: Loan term, number of months
          example: 36
        firstName:
          type: string
          pattern: "^[a-zA-Zа-яёА-ЯЁ-]{1,60}$"
          description: First name of client
          example: "Ivan"
        lastName:
          type: string
          pattern: "^[a-zA-Zа-яёА-ЯЁ-]{1,60}$"
          description: Last name of client
          example: "Ivanov"
        middleName:
          type: string
          pattern: "^$|^[a-zA-Zа-яёА-ЯЁ-]{1,60}$"
          description: Middle name of client
          example: "Ivanovich"
        email:
          type: string
          pattern: "^[a-z0-9A-Z_!#$%&'*+/=?`{|}~^.-]+@[a-z0-9A-Z.-]+$"
          description: Email of client
          example: "ivan@mail.ru"
        birthDate:
          type: string
          format: date
          description: Date of birth of client
          example: "2000-01-01"
        passportSeries:
          type: string
          pattern: "^[0-9]{4}$"
          description: Passport series of client
          example: "4567"
        passportNumber:
          type: string
          pattern: "^[0-9]{6}$"
          description: Passport number of client
          example: "578430"

    LoanOfferDto:
      type: object
      description: Information about loan offer
      required:
        - statementId
        - requestedAmount
        - totalAmount
        - term
        - monthlyPayment
        - rate
        - isInsuranceEnabled
        - isSalaryClient
      properties:
        statementId:
          type: string
          format: uuid
          description: Unique id of loan statement
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        requestedAmount:
          type: number
          format: bigdecimal
          description: Requested loan amount
          example: 100000
        totalAmount:
          type: number
          format: bigdecimal
          description: Loan amount including interest and insurance costs
          example: 130000
        term:
          type: integer
          description: Loan term, number of months
          example: 36
        monthlyPayment:
          type: number
          format: bigdecimal
          description: The amount of payment for every month
          example: 10000
        rate:
          type: number
          format: bigdecimal
          description: Loan rate calculated depending on parameters, percentages
          example: 22
        isInsuranceEnabled:
          type: boolean
          description: Information about whether insurance is included
          example: true
        isSalaryClient:
          type: boolean
          description: Information about whether the client is a salary client
          example: false

    FinishRegistrationRequestDto:
      type: object
      description: Information about loan offer
      required:
        - gender
        - maritalStatus
        - dependentAmount
        - passportIssueDate
        - passportIssueBranch
        - employment
        - accountNumber
      properties:
        gender:
          type: string
          description: Gender of client (MALE, FEMALE, NON_BINARY)
          example: "MALE"
        maritalStatus:
          type: string
          description: Marital status of client (MARRIED, SINGLE, DIVORCED, WIDOW_WIDOWER)
          example: "SINGLE"
        dependentAmount:
          type: integer
          description: Number of client's dependents
          example: 0
        passportIssueDate:
          type: string
          format: date
          description: Date of issue of the client's passport
          example: "2021-01-01"
        passportIssueBranch:
          type: string
          description: Client's passport issuing department
          example: "Passport issuance department of the federal service for the saratov region"
        employment:
          $ref: '#/components/schemas/EmploymentDto'
        accountNumber:
          type: string
          description: Client's account number
          example: "4654335"

    EmploymentDto:
      type: object
      description: Detailed information about employment
      required:
        - employmentStatus
        - employerINN
        - salary
        - position
        - workExperienceTotal
        - workExperienceCurrent
      properties:
        employmentStatus:
          type: string
          description: Employment status of client (UNEMPLOYED, SELF_EMPLOYED, EMPLOYED, BUSINESS_OWNER)
          example: "EMPLOYED"
        employerINN:
          type: string
          description: Taxpayer identification number of client's employer
          example: "520205004556"
        salary:
          type: number
          format: bigdecimal
          description: Size of client's salary
          example: 40000
        position:
          type: string
          description: Position at work (WORKER, MID_MANAGER, TOP_MANAGER, OWNER)
          example: "WORKER"
        workExperienceTotal:
          type: integer
          description: Total work experience, number of months
          example: 36
        workExperienceCurrent:
          type: integer
          description: Current work experience, number of months
          example: 24

    StatementDto:
      type: object
      description: Information about statement
      required:
        - statementId
        - clientId
        - creditId
        - status
        - creationDate
        - appliedOffer
        - signDate
        - sesCode
        - statusHistory
      properties:
        statementId:
          type: string
          description: Id of statement
          example: "3fa85f64-5717-4562-b3fc-2c963f66afa6"
        clientId:
          type: string
          description: Id of client
          example: "4fa85f64-5717-4562-b3fc-2c963f66afa4"
        creditId:
          type: string
          description: Id of credit
          example: "5fa85f64-5717-4562-b3fc-2c963f66afa2"
        status:
          type: string
          description: Status of statement (PREAPPROVAL, APPROVED, CC_DENIED, CC_APPROVED, PREPARE_DOCUMENTS,DOCUMENT_CREATED, CLIENT_DENIED, DOCUMENT_SIGNED, CREDIT_ISSUED)
          example: "APPROVED"
        creationDate:
          type: string
          format: date-time
          description: Statement creation time
          example: "2025-07-07 18:53:42.951815"
        appliedOffer:
          $ref: '#/components/schemas/LoanOfferDto'
        signDate:
          type: string
          format: date-time
          description: Document signing time
          example: "2025-07-08 18:53:42.951815"
        sesCode:
          type: string
          description: Code for signing document
          example: "4J5CK3"
        statusHistory:
          type: array
          items:
            $ref: '#/components/schemas/StatementStatusHistoryDto'
          example:
            - number: 1
              status: "APPROVED"
              time: "2025-07-07 18:53:42.951815"
              changeType: "AUTOMATIC"
            - number: 2
              status: "CC_DENIED"
              time: "2025-07-07 18:54:42.951815"
              changeType: "AUTOMATIC"

    StatementStatusHistoryDto:
      type: object
      description: Information about changing the statement
      required:
        - status
        - time
        - changeType
      properties:
        status:
          type: string
          description: Status of statement (PREAPPROVAL, APPROVED, CC_DENIED, CC_APPROVED, PREPARE_DOCUMENTS,DOCUMENT_CREATED, CLIENT_DENIED, DOCUMENT_SIGNED, CREDIT_ISSUED)
          example: "APPROVED"
        time:
          type: string
          format: date-time
          description: Statement change time
          example: "2025-07-07T18:53:42.951"
        changeType:
          type: string
          description: Change type of statement (AUTOMATIC, MANUAL)
          example: "AUTOMATIC"